Keep shortcut settings changes within CI budgets
The PR branch now rebases onto current main, whose Swift file budget is tighter around cmuxApp and AppDelegate. Move the shortcut settings row/control views out of cmuxApp and keep the unbound numbered-shortcut check on the existing branch so the behavior remains unchanged while the budget guard has room again.

Constraint: CI rejects growth in tracked large Swift files and warning budgets.

Rejected: Refresh the file budget | would hide avoidable large-file growth.

Confidence: high

Scope-risk: narrow

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Tested: git diff --check

Tested: plutil -lint GhosttyTabs.xcodeproj/project.pbxproj && jq empty Resources/Localizable.xcstrings

Not-tested: Local test suites per repository policy.
